Extracting ListFilter to a React component
